Address simplify review: reuse NavigateToTabAndFind, tighten comments
- Promote NavigateToTabAndFind from TextAndValueTests to TestAppFixture
  so ClickTests can reuse it instead of reimplementing tab navigation
  inline. Update the existing TextAndValueTests call sites.
- Tighten the WHAT-comments in both ClickTool.cs and the new modal-hang
  test down to a single WHY sentence each.
- Inline the gratuitous invokePattern local in ClickTool — closure
  capture already keeps the element reference alive.
- Replace the dual-timeout assertions in the modal-hang test with a
  single 2s WhenAny race; the second elapsed-ms check was redundant.
