+++ tests/FlaUI.Mcp.IntegrationTests/TextAndValueTests.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using PlaywrightWindows.Mcp.Tools;
 using Xunit.Abstractions;
 
@@ -19,35 +18,11 @@ public class TextAndValueTests
         _output = output;
     }
 
-    /// <summary>
-    /// Navigate to a tab and poll until an expected element appears.
-    /// Returns the ref of the expected element.
-    /// </summary>
-    private async Task<string> NavigateToTabAndFind(string windowHandle, string tabName, string elementName)
-    {
-        var tabRef = _fixture.FindRefByName(windowHandle, tabName);
-        Assert.NotNull(tabRef);
-
-        var clickTool = new ClickTool(_fixture.Elements);
-        await _fixture.CallTool(clickTool, new { @ref = tabRef });
-
-        // Poll for the element to appear after tab switch
-        var sw = Stopwatch.StartNew();
-        while (sw.ElapsedMilliseconds < 5000)
-        {
-            await Task.Delay(100);
-            var found = _fixture.FindRefByName(windowHandle, elementName);
-            if (found != null) return found;
-        }
-
-        Assert.Fail($"Element \"{elementName}\" not found after navigating to \"{tabName}\" tab.");
-        return ""; // unreachable
-    }
 
     [Fact]
     public async Task WinForms_GetText_ReadOnlyField()
     {
-        var resultRef = await NavigateToTabAndFind(_fixture.WinFormsHandle, "Forms", "Result");
+        var resultRef = await _fixture.NavigateToTabAndFind(_fixture.WinFormsHandle, "Forms", "Result");
 
         var tool = new GetTextTool(_fixture.Elements);
         var text = await _fixture.CallTool(tool, new { @ref = resultRef });
@@ -58,7 +33,7 @@ public class TextAndValueTests
     [Fact]
     public async Task WinForms_TypeAndGetText()
     {
-        var nameRef = await NavigateToTabAndFind(_fixture.WinFormsHandle, "Forms", "Name");
+        var nameRef = await _fixture.NavigateToTabAndFind(_fixture.WinFormsHandle, "Forms", "Name");
 
         // Fill the name field
         var fillTool = new FillTool(_fixture.Elements);
@@ -78,7 +53,7 @@ public class TextAndValueTests
     [Fact]
     public async Task Wpf_GetText_ReadOnlyField()
     {
-        var resultRef = await NavigateToTabAndFind(_fixture.WpfHandle, "Forms", "Result");
+        var resultRef = await _fixture.NavigateToTabAndFind(_fixture.WpfHandle, "Forms", "Result");
 
         var tool = new GetTextTool(_fixture.Elements);
         var text = await _fixture.CallTool(tool, new { @ref = resultRef });
